feat: :sparkles: Move objectset func from Parse to Scene folder
오브젝트 세팅 파일 파서 -> 씬 폴더로 수정, header, makefile 수정

#27
