+++ src/scene/ambient.c
