+++ src/scene/light.c
