+++ src/scene/objects.c
